// frontend/src/models/SampleData.ts
export interface SampleData {
  registrationNo: string;
  batchNo: string;
  sampleName: string;
  sampleCode: string;
  lab: string;
  parameter: string;
  paraCode: string;
  methodName: string;
  methodCode: string;
  registrationDate: string;
  recieptDate: string;
  analysisStartDate: string;
  analysisCompletionDate: string;
  tatDate: string;
}